+++ service/index.js
@@ -1,6 +1,9 @@
 const hrtime = require('browser-process-hrtime');
 const utport = require('ut-port');
 const path = require('path');
+const {version} = require('../package.json');
+const gte = require('semver/functions/gte');
+const joi = require('joi');
 
 module.exports = ({serviceBus, logFactory, log, vfs}) => {
     const watch = (filename, fn) => {
@@ -26,7 +29,7 @@ module.exports = ({serviceBus, logFactory, log, vfs}) => {
         });
     };
 
-    const servicePorts = utport.ports({bus: serviceBus.publicApi, logFactory, vfs});
+    const servicePorts = utport.ports({bus: serviceBus.publicApi, logFactory, vfs, joi, version: wanted => gte(version, wanted)});
 
     function configure(obj = {}, config, moduleName, pkg) {
         return [].concat(...Object.entries(obj).map(([name, value]) => {
@@ -130,13 +133,29 @@ module.exports = ({serviceBus, logFactory, log, vfs}) => {
         return prev;
     }, []);
 
-    const create = async(serviceConfig, config, test) => servicePorts.create(
-        await load(serviceConfig, config, test),
-        config || {},
-        test
-    );
+    const create = async(serviceConfig, config, test) => {
+        try {
+            return await servicePorts.create(
+                await load(serviceConfig, config, test),
+                config || {},
+                test
+            );
+        } catch (error) {
+            if (!error.type) error.type = 'serviceLayer.create';
+            log.error && log.error(error);
+            throw new Error('silent');
+        }
+    };
 
-    const start = ports => servicePorts.start(ports);
+    const start = async ports => {
+        try {
+            return await servicePorts.start(ports);
+        } catch (error) {
+            if (!error.type) error.type = 'serviceLayer.start';
+            log.error && log.error(error);
+            throw new Error('silent');
+        }
+    };
 
     const method = fn => ({filter, ...rest} = {}) => {
         const portsAndModules = servicePorts.fetch(filter);